requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
typer[all]>=0.9.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.0
numpy>=1.26.0
scikit-learn>=1.3.0
hdbscan>=0.8.33
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
httpx>=0.25.0
aiofiles>=23.2.0
datasketch>=1.6.0
